--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,7 @@ lazy_static = "1.4.0"
 sexp = "1.1.4"
 itertools = "0.10.3"
 rand = "0.8.4"
+parking_lot = "0.12.0"
 
 [profile.release]
 debug = true # for flamegraphs