Adds multithreading. Already seems to work.


Guiding idea: we don't want to make any changes that will make it harder to implement new features! That's not worth it. We just want some basic bare bones parallelism that might not be incredibly fast but will give some boost.

Why: if we're trying to do a performance comparison ofc doing some parallelizing makes sense!

Downsides:
* we need to commit to making everything use `Arc` instead of normal references
* who knows if the thread boundary will get annoying when adding future features. My instinct is given that no real issues came up during this implementation, this might actually be a good move to use multithreading. Thankfully this change only needs to affect things starting at `derive_inventions` as we can just convert everything to Arc at that point.

Performance: Using `--threads=1` (the default) seems to give the same performance as usual, while adding more threads  pretty rapidly gives better performance (eg 5x, 10x). `dials.json` drops from 75s to 9s for me (arity 2, 3 iterations) and `nuts-bolts.json` drops from 7s to 1.5s for me. I'm guessing the larger the workloads the more we'll be able to take advantage of threads. Overshooting eg `--threads=30` doesn't seem to do any worse for me though it's not better than using less. Note that along the way I tested the little changes I made to make things multithreading-compatible and they didnt slow down the single theraded case noticably.

I chose to use `parking_lot` as the mutex crate, though it seems at least on my machine that didn't speed anything up. My impression is that the standard library can be slow on some machines though so I think it makes sense to use `parking_lot` (a crate that they'll probably merge into std at some point).

Todos:
- [ ] clean up and comment things, I did this pretty quick
- [ ] decide how to handle Stats. Probably we should have a --stats option to turn it on, and when it's on then threads will need to do mutex unlocks to modify it. Also have it default to be turned on in the single threaded case since it wont slow it down anyways (itll still have to grab the lock but itll have no competition). Probably even in the competitive case the lock will be grabbed for so little time that it wont be a major slowdown.